+++ HTTPLayer/CHANGELOG.md
@@ -9,4 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GET requests with PathParameters
 - GET requests with QueryParameters
 - POST requests with ObjectParameter
-- PUT request with ObjectParameter
+- PUT request with ObjectParameter and/or PathParameters
+- PUT request with ObjectParameter and/or QueryParameters
+- DELETE request with ObjectParameter and/or PathParameters
+- DELETE request with ObjectParameter and/or QueryParameters